Rename: states/authState -> authStore
