+++ src/components/FollowingPage/RecommendCard/RecommendCard.tsx
@@ -8,7 +8,7 @@ import { useToast } from "@/hooks/useToast";
 import { putFollow } from "@/api/users/putIdFollow";
 import { deleteFollow } from "@/api/users/deleteIdFollow";
 import IconComponent from "@/components/Asset/Icon";
-import { useAuthStore } from "@/states/authState";
+import { useAuthStore } from "@/states/authStore";
 import { usePreventRightClick } from "@/hooks/usePreventRightClick";
 
 export default function RecommendCard({
