+++ src/components/PopularPage/PopularFeed/FeedCard/FeedCard.tsx
@@ -3,7 +3,7 @@ import IconComponent from "@/components/Asset/Icon";
 import styles from "./FeedCard.module.scss";
 import { formatCurrency } from "@/utils/formatCurrency";
 import Link from "next/link";
-import { useAuthStore } from "@/states/authState";
+import { useAuthStore } from "@/states/authStore";
 import { deleteLike, putLike } from "@/api/feeds/putDeleteFeedsLike";
 import { FeedCardProps } from "./FeedCard.types";
 import { usePreventRightClick } from "@/hooks/usePreventRightClick";
